core/python/requirements.txt
# For Jupyter notebook
pyzmq == 16.0.2
tornado == 4.4.2

# For embedded node
py-mini-racer == 0.1.8

tensorflow == 1.1.0rc2
